move incorrectly placed specfile
